update: separated signout component
